api: create project with some settings
This commit adds the use of the docker, which helps in the development of the
project. You can run this command to start the project:

```
- docker-compose up
```

With that, all project dependencies will be downloaded and installed inside the
container. In addition, the nodemon package is used to refresh the project, and
the eslint airbnb package is installed for the project to follow a pattern.